backend/requirements.txt
# Rural AI Assistant Backend — production dependencies

# Framework
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0

# Database
sqlalchemy>=2.0.0,<3.0
psycopg2-binary>=2.9.9,<3.0

# AWS
boto3>=1.34.0,<2.0

# Config & validation
python-dotenv>=1.0.0,<2.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0

# Optional for future use (uncomment when needed):
# twilio>=8.0.0
# google-cloud-speech>=2.24.0
# boto3  # Polly is part of boto3
# langgraph
# pinecone-client
